pyproject.toml
[project]
name = "questions-vote"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.6",
    "requests>=2.32.3",
    "ruff>=0.9.3",
]

[tool.ruff]
line-length = 120
